+++ dhcp-clients-webapp-backend/pkg/uibackend/doc.go
@@ -1,0 +1,3 @@
+// This package implements an HTML/Websocket server that provides a tabular view of
+// all DHCP clients served by the dnsmasq instance that lives into this HomeAssistant addon
+package uibackend
