+++ dhcp-clients-webapp-backend/pkg/uibackend/uibackend.go
@@ -1,5 +1,3 @@
-// This package implements an HTML/Websocket server that provides a tabular view of
-// all DHCP clients served by the dnsmasq instance that lives into this HomeAssistant addon
 package uibackend
 
 import (
@@ -15,7 +13,6 @@ import (
 	"os"
 	"slices"
 	"sync"
-	"time"
 
 	"github.com/b0ch3nski/go-dnsmasq-utils/dnsmasq"
 	"github.com/gorilla/websocket"
@@ -29,99 +26,6 @@ var dnsmasqMarkerForMissingHostname = "*"
 // These absolute paths must be in sync with the Dockerfile
 var staticWebFilesDir = "/opt/web/static"
 var templatesDir = "/opt/web/templates"
-
-// DhcpClientData holds all the information the backend has about a particular DHCP client
-type DhcpClientData struct {
-	// the lease as it is parsed from dnsmasq LEASE file:
-	Lease dnsmasq.Lease
-
-	// metadata associated with the DHCP client (obtained from configuration):
-
-	// HasStaticIP indicates whether the DHCP server is configured to provide a specific IP address
-	// (i.e. has an IP address reservation) for this client.
-	// Note that static IP addresses do not need to be inside the DHCP range; indeed quite often the
-	// static IP address reserved lies outside the DHCP range
-	HasStaticIP bool
-
-	// IsInsideDHCPPool indicates whether this DHCP client has an IP that lies within the DHCP pool
-	// range and thus is consuming an IP address from that pool
-	// (note that this DHCP client might be a client with a static reservation or not)
-	IsInsideDHCPPool bool
-
-	// Sometimes the hostname provided by the DHCP client to the DHCP server is really awkward and
-	// non-informative, so we allow users to override that.
-	FriendlyName string
-}
-
-func LeaseTimeToString(t time.Time) string {
-
-	if t.IsZero() {
-		return "Never expires"
-	}
-
-	now := time.Now()
-	duration := t.Sub(now)
-	if duration < 0 {
-		return "Expired"
-	}
-
-	// compute hours, min, secs
-	days := int(duration.Hours()) / 24
-	hours := int(duration.Hours()) % 24
-	minutes := int(duration.Minutes()) % 60
-	seconds := int(duration.Seconds()) % 60
-
-	if days > 0 {
-		return fmt.Sprintf("%02dd, %02dh, %02dm, %02ds", days, hours, minutes, seconds)
-	} else if hours > 0 {
-		return fmt.Sprintf("%02dh, %02dm, %02ds", hours, minutes, seconds)
-	} else {
-		return fmt.Sprintf("%02dm, %02ds", minutes, seconds)
-	}
-}
-
-// MarshalJSON customizes the JSON serialization for DhcpClientData
-func (d DhcpClientData) MarshalJSON() ([]byte, error) {
-	return json.Marshal(&struct {
-		Lease struct {
-			Expires  int64  `json:"expires"`
-			MacAddr  string `json:"mac_addr"`
-			IPAddr   string `json:"ip_addr"`
-			Hostname string `json:"hostname"`
-		} `json:"lease"`
-		HasStaticIP      bool   `json:"has_static_ip"`
-		IsInsideDHCPPool bool   `json:"is_inside_dhcp_pool"`
-		FriendlyName     string `json:"friendly_name"`
-	}{
-		Lease: struct {
-			Expires  int64  `json:"expires"`
-			MacAddr  string `json:"mac_addr"`
-			IPAddr   string `json:"ip_addr"`
-			Hostname string `json:"hostname"`
-		}{
-			Expires:  d.Lease.Expires.Unix(), // unix time, the number of seconds elapsed since January 1, 1970 UTC
-			MacAddr:  d.Lease.MacAddr.String(),
-			IPAddr:   d.Lease.IPAddr.String(),
-			Hostname: d.Lease.Hostname,
-		},
-		HasStaticIP:      d.HasStaticIP,
-		IsInsideDHCPPool: d.IsInsideDHCPPool,
-		FriendlyName:     d.FriendlyName,
-	})
-}
-
-// DhcpClientFriendlyName is the 1:1 binding between a MAC address and a human-friendly name
-type DhcpClientFriendlyName struct {
-	MacAddress   net.HardwareAddr
-	FriendlyName string
-}
-
-// IpAddressReservation represents a static IP configuration loaded from the addon configuration file
-type IpAddressReservation struct {
-	Name string `json:"name"`
-	Mac  string `json:"mac"`
-	IP   string `json:"ip"`
-}
 
 type UIBackend struct {
 	// Static IP addresses, as read from the configuration
@@ -406,26 +310,6 @@ func (b *UIBackend) readCurrentLeaseFile() error {
 	}
 
 	return b.processLeaseUpdatesFromArray(leases)
-}
-
-// dummy struct used to unmarshal HomeAssistant option file correctly
-type AddonConfig struct {
-	IpAddressReservations []IpAddressReservation `json:"ip_address_reservations"`
-
-	DhcpClientsFriendlyNames []struct {
-		Name string `json:"name"`
-		Mac  string `json:"mac"`
-	} `json:"dhcp_clients_friendly_names"`
-
-	DhcpRange struct {
-		StartIP string `json:"start_ip"`
-		EndIP   string `json:"end_ip"`
-	} `json:"dhcp_range"`
-
-	LogDHCP  bool `json:"log_dhcp"`
-	LogWebUI bool `json:"log_web_ui"`
-
-	WebUIPort int `json:"web_ui_port"`
 }
 
 func (b *UIBackend) readAddonConfig() error {
